requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
jinja2>=3.1.4
python-dotenv>=1.0.1
httpx>=0.27.0

# Database
sqlalchemy[asyncio]>=2.0.0
asyncpg>=0.29.0
alembic>=1.13.0

# RSS ingestion
feedparser>=6.0.11

# OpenSearch
opensearch-py[async]>=2.4.0

# Auth
bcrypt>=4.0.0
python-jose[cryptography]>=3.3.0
python-multipart>=0.0.9
email-validator>=2.0.0
